Get back to directory where script is executed
The script `docker-etnrypoint.sh` is designed so as to be executed
from a shell in a running container. When an error happens while
building the Grasp Plugin the shell catches the error and exits
immediately. In such a case, the working directory remains in the
`build` directory. This commit fixes this by getting back to the
directory where the script starts even when an error occurs.